MOVE: B64 to utils
