+++ src/lib/kotlin/slatekit-app/src/main/kotlin/slatekit/app/AppContext.kt
@@ -17,7 +17,7 @@ import slatekit.common.Context
 import slatekit.common.args.Args
 import slatekit.common.conf.Conf
 import slatekit.common.conf.Config
-import slatekit.common.encrypt.B64Java8
+import slatekit.common.utils.B64Java8
 import slatekit.common.encrypt.Encryptor
 import slatekit.common.envs.Env
 import slatekit.common.envs.EnvMode
