+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/Defaults.kt
@@ -3,7 +3,7 @@ package slatekit.common
 import slatekit.common.args.Args
 import slatekit.common.conf.Conf
 import slatekit.common.conf.Config
-import slatekit.common.encrypt.B64Java8
+import slatekit.common.utils.B64Java8
 import slatekit.common.encrypt.Encryptor
 import slatekit.common.envs.Env
 import slatekit.common.envs.EnvMode
