+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/encrypt/Encryptor.kt
@@ -13,6 +13,8 @@
 
 package slatekit.common.encrypt
 
+import slatekit.common.utils.B64
+
 /**
  * Provides a simple facade for encryption using AES
  * @param key : The secret key ( must be 16, or 32 bytes long )
@@ -20,7 +22,7 @@ package slatekit.common.encrypt
  */
 open class Encryptor(private val key: String,
                      private val iv: String,
-                     private val b64:B64) {
+                     private val b64: B64) {
 
     fun encrypt(text: String): String = EncryptorAES.encrypt(key, iv, text, b64)
