+++ src/lib/kotlin/slatekit-examples/src/main/kotlin/slatekit/examples/Example_Config.kt
@@ -20,7 +20,7 @@ import slatekit.results.Success
 import slatekit.common.conf.ConfFuncs
 import slatekit.common.conf.Config
 import slatekit.common.db.DbCon
-import slatekit.common.encrypt.B64Java8
+import slatekit.common.utils.B64Java8
 import slatekit.common.encrypt.Encryptor
 import slatekit.cmds.Command
 import slatekit.cmds.CommandRequest
