+++ src/lib/kotlin/slatekit-examples/src/main/kotlin/slatekit/examples/Example_Context.kt
@@ -21,7 +21,7 @@ import slatekit.common.CommonContext
 import slatekit.common.args.Args
 import slatekit.common.args.ArgsSchema
 import slatekit.common.conf.Config
-import slatekit.common.encrypt.B64Java8
+import slatekit.common.utils.B64Java8
 import slatekit.common.encrypt.Encryptor
 import slatekit.common.envs.Env
 import slatekit.common.envs.EnvMode
