+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/setup/MyEncryptor.kt
@@ -12,7 +12,7 @@
  */
 package test.setup
 
-import slatekit.common.encrypt.B64Java8
+import slatekit.common.utils.B64Java8
 import slatekit.common.encrypt.Encryptor
 
 object MyEncryptor : Encryptor("wejklhviuxywehjk", "3214maslkdf03292", B64Java8) {
