+++ src/lib/kotlin/slatekit/src/main/kotlin/slatekit/samples/app/App.kt
@@ -4,7 +4,7 @@ import slatekit.app.App
 import slatekit.app.AppOptions
 import slatekit.common.Context
 import slatekit.common.args.ArgsSchema
-import slatekit.common.encrypt.B64Java8
+import slatekit.common.utils.B64Java8
 import slatekit.common.encrypt.Encryptor
 import slatekit.common.info.About
 import slatekit.results.Success
